Moved Auth Pages to views directory
